refactor: 중복을 고려한 자동차 id 자료형 변경 (int -> long)

+++ src/main/java/racingcar/domain/Car.java
@@ -4,9 +4,9 @@ import racingcar.vo.CarName;
 
 public class Car {
 
-    private static int count = 0; // ID 를 위한 정적 변수 (생성자를 호출할 때 마다, 카운트 증가 → 각각의 차 ID)
+    private static long count = 0L; // ID 를 위한 정적 변수 (생성자를 호출할 때 마다, 카운트 증가 → 각각의 자동차 ID)
 
-    private final int id; // 중복 이름을 고려한, 차 ID
+    private final long id; // 중복 이름을 고려한, 자동차 ID
     private final CarName carName;
     private int forwardCount;
